+++ Questionable/Windows/ConfigComponents/GeneralConfigComponent.cs
@@ -96,6 +96,8 @@ internal sealed class GeneralConfigComponent : ConfigComponent
         using ImRaii.TabItemDisposable tab = ImRaii.TabItem(_L("General") + "###General");
         if (!tab)
             return;
+        var size = ImGui.GetWindowContentRegionMax();
+        using var _ = ImRaii.TextWrapPos(size.X);
         Dictionary<string, string> languages = new(StringComparer.Ordinal){
             { "en",    _L("English") },
             { "ja-jp", _L("Japanese") },
@@ -129,6 +131,7 @@ internal sealed class GeneralConfigComponent : ConfigComponent
         if (ImGui.CollapsingHeader(_L("Preferences")))
         {
             ECombatModule combatModule = Configuration.General.CombatModule;
+            ImGui.SetNextItemWidth(size.X / 2);
             if (ImGuiEx.EnumCombo(_L("Preferred Combat Module"), ref combatModule))
             {
                 Configuration.General.CombatModule = combatModule;
@@ -137,6 +140,7 @@ internal sealed class GeneralConfigComponent : ConfigComponent
 
             (uint[] mountIds, string[] mountNames) = _mounts.Value;
             uint mountId = Configuration.General.MountId;
+            ImGui.SetNextItemWidth(size.X / 2);
             if (ImGuiComponentsLocal.DrawSearchableCombo(_L("Preferred Mount"), mountIds, mountNames,
                 Configuration.General.MountId, ref _mountSearchString, ref mountId))
             {
@@ -145,6 +149,7 @@ internal sealed class GeneralConfigComponent : ConfigComponent
             }
 
             int grandCompany = (int)Configuration.General.GrandCompany;
+            ImGui.SetNextItemWidth(size.X / 2);
             if (ImGui.Combo(_L("Preferred Grand Company"), ref grandCompany, _grandCompanyNames,
                 _grandCompanyNames.Length))
             {
@@ -153,23 +158,28 @@ internal sealed class GeneralConfigComponent : ConfigComponent
             }
 
             (Job[] classJobIds, string[] classJobNames) = _classJobs.Value;
+            ImGui.SetNextItemWidth(size.X / 2);
             DrawComboOption(_L("Preferred Combat Job"), classJobIds, classJobNames,
                 () => Configuration.General.CombatJob,
                 v => Configuration.General.CombatJob = v);
 
             (Job[] craftJobIds, string[] craftJobNames) = _craftJobs.Value;
+            ImGui.SetNextItemWidth(size.X / 2);
             DrawComboOption(_L("Preferred Crafting Job"), craftJobIds, craftJobNames,
                 () => Configuration.General.CraftingJob,
                 v => Configuration.General.CraftingJob = v);
 
             (Job[] gatherJobIds, string[] gatherJobNames) = _gatherJobs.Value;
+            ImGui.SetNextItemWidth(size.X / 2);
             DrawComboOption(_L("Preferred Gathering Job"), gatherJobIds, gatherJobNames,
                 () => Configuration.General.GatheringJob,
                 v => Configuration.General.GatheringJob = v);
 
+            ImGui.BeginGroup();
             using (ImRaii.Disabled(!StylistIpc.IsInstalled))
             {
                 EGearsetUpdateSource gearsetSource = Configuration.General.GearsetUpdateSource;
+                ImGui.SetNextItemWidth(size.X / 2);
                 if (ImGuiEx.EnumCombo(_L("Preferred Gear Upgrade Source"), ref gearsetSource))
                 {
                     Configuration.General.GearsetUpdateSource = gearsetSource;
@@ -182,8 +192,12 @@ internal sealed class GeneralConfigComponent : ConfigComponent
                     Save();
                 }
             }
+            ImGui.EndGroup();
+            if (!StylistIpc.IsInstalled && ImGui.IsItemHovered())
+                ImGui.SetTooltip(_L("Stylist is not installed."));
 
             string chocoboName = Configuration.General.ChocoboName;
+            ImGui.SetNextItemWidth(size.X / 2);
             if (ImGui.InputText(_L("Chocobo name"), ref chocoboName, 20))
                 Configuration.General.ChocoboName = chocoboName;
 
@@ -204,6 +218,7 @@ internal sealed class GeneralConfigComponent : ConfigComponent
             }
 
             string displayName = Configuration.General.DisplayName;
+            ImGui.SetNextItemWidth(size.X / 2);
             if (ImGui.InputText(_L("Display name"), ref displayName, 20))
                 Configuration.General.DisplayName = displayName;