+++ Questionable/Windows/ConfigComponents/StopConditionComponent.cs
@@ -28,6 +28,7 @@ internal sealed class StopConditionComponent : ConfigComponent
     private readonly QuestSelector _completeQuestSelector;
     private readonly QuestTooltipComponent _questTooltipComponent;
     private readonly UiUtils _uiUtils;
+    private Vector2 Size;
 
     public StopConditionComponent(
         IDalamudPluginInterface pluginInterface,
@@ -74,6 +75,8 @@ internal sealed class StopConditionComponent : ConfigComponent
         using ImRaii.TabItemDisposable tab = ImRaii.TabItem(_L("Stop") + "###StopConditionns");
         if (!tab)
             return;
+        Size = ImGui.GetWindowContentRegionMax();
+        var wrap = ImRaii.TextWrapPos(Size.X+10);
 
         bool runCommand = Configuration.Stop.RunCommandAfterStop;
         if (ImGui.Checkbox(_L("Run command when Questionable finishes automatic questing"), ref runCommand))
@@ -81,8 +84,6 @@ internal sealed class StopConditionComponent : ConfigComponent
             Configuration.Stop.RunCommandAfterStop = runCommand;
             Save();
         }
-        ImGui.SameLine();
-        ImGui.TextColored(ImGuiColors.DalamudRed, _L("Experimental feature"));
         string command = Configuration.Stop.CommandAfterStop;
         if (ImGui.InputText(_L("Command"), ref command, 128))
         {